Move seeds to subdirectories
