+++ db/seeds/exercises/chest.rb
@@ -1,26 +1,28 @@
 # frozen_string_literal: true
 
-module Exercises
-  class Chest
-    CHEST = [
-      'Barbell Inclide Bench Press',
-      'Bench Press',
-      'Butterfly',
-      'Cable Crossover',
-      'Decline Dumbbell Bench Press',
-      'Dumbbell Bench Press',
-      'Dumbbell Incline Bench Press',
-      'Dumbbell Pullover',
-      'Incline Dumbbell Flyes',
-      'Smith Machine Bench Press'
-    ].freeze
+module Seeds
+  module Exercises
+    class Chest
+      CHEST = [
+        'Barbell Inclide Bench Press',
+        'Bench Press',
+        'Butterfly',
+        'Cable Crossover',
+        'Decline Dumbbell Bench Press',
+        'Dumbbell Bench Press',
+        'Dumbbell Incline Bench Press',
+        'Dumbbell Pullover',
+        'Incline Dumbbell Flyes',
+        'Smith Machine Bench Press'
+      ].freeze
 
-    def params
-      {
-        training_type: :strength_training,
-        muscle_group: :chest,
-        items: CHEST
-      }
+      def params
+        {
+          training_type: :strength_training,
+          muscle_group: :chest,
+          items: CHEST
+        }
+      end
     end
   end
 end
