+++ db/seeds/exercises/shoulders.rb
@@ -1,26 +1,28 @@
 # frozen_string_literal: true
 
-module Exercises
-  class Shoulders
-    SHOULDERS = [
-      'Arnold Dumbbell Press',
-      'Barbell Upright Row',
-      'Bent Over Low-Pulley Side Lateral',
-      'Bent Over Rear Delt Row With Head on Bench',
-      'Cable Seated Rear Lateral Raise',
-      'Dumbbell Shoulder Press',
-      'Front Cable Raise',
-      'Machine Shoulder (Military) Press',
-      'Seated Barbell Military Press',
-      'Side Lateral Raise'
-    ].freeze
+module Seeds
+  module Exercises
+    class Shoulders
+      SHOULDERS = [
+        'Arnold Dumbbell Press',
+        'Barbell Upright Row',
+        'Bent Over Low-Pulley Side Lateral',
+        'Bent Over Rear Delt Row With Head on Bench',
+        'Cable Seated Rear Lateral Raise',
+        'Dumbbell Shoulder Press',
+        'Front Cable Raise',
+        'Machine Shoulder (Military) Press',
+        'Seated Barbell Military Press',
+        'Side Lateral Raise'
+      ].freeze
 
-    def params
-      {
-        training_type: :strength_training,
-        muscle_group: :shoulders,
-        items: SHOULDERS
-      }
+      def params
+        {
+          training_type: :strength_training,
+          muscle_group: :shoulders,
+          items: SHOULDERS
+        }
+      end
     end
   end
 end
